SignalR generation in dotnet core version
